add recommendation table
